+++ crates/github_scbot_database/src/schema.rs
@@ -73,6 +73,7 @@ table! {
         default_needed_reviewers_count -> Int4,
         pr_title_validation_regex -> Text,
         manual_interaction -> Bool,
+        default_automerge -> Bool,
     }
 }