+++ migrations/2021-09-20-094351_default-automerge/down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE repository DROP COLUMN default_automerge;